Transform copy of libjpeg-turbo to module
