+++ lib/jpegli/common.h
@@ -20,7 +20,10 @@
 #ifndef LIB_JPEGLI_COMMON_H_
 #define LIB_JPEGLI_COMMON_H_
 
+/* clang-format off */
+#include <stdio.h>
 #include <jpeglib.h>
+/* clang-format on */
 
 #include "lib/jpegli/types.h"
