+++ lib/jpegli/decode_marker.h
@@ -6,8 +6,9 @@
 #ifndef LIB_JPEGLI_DECODE_MARKER_H_
 #define LIB_JPEGLI_DECODE_MARKER_H_
 
-#include <jpeglib.h>
 #include <stdint.h>
+
+#include "lib/jpegli/common.h"
 
 namespace jpegli {
