+++ lib/jpegli/encode_internal.h
@@ -6,10 +6,10 @@
 #ifndef LIB_JPEGLI_ENCODE_INTERNAL_H_
 #define LIB_JPEGLI_ENCODE_INTERNAL_H_
 
-#include <jpeglib.h>
 #include <stdint.h>
 
 #include "lib/jpegli/bit_writer.h"
+#include "lib/jpegli/common.h"
 #include "lib/jpegli/common_internal.h"
 #include "lib/jpegli/encode.h"
