+++ lib/jpegli/entropy_coding-inl.h
@@ -37,76 +37,6 @@ using hwy::HWY_NAMESPACE::Sub;
 using DI = HWY_FULL(int32_t);
 constexpr DI di;
 
-void ZigZagShuffle(int32_t* JXL_RESTRICT block) {
-  // TODO(szabadka) SIMDify this.
-  int32_t tmp[DCTSIZE2];
-  tmp[0] = block[0];
-  tmp[1] = block[1];
-  tmp[2] = block[8];
-  tmp[3] = block[16];
-  tmp[4] = block[9];
-  tmp[5] = block[2];
-  tmp[6] = block[3];
-  tmp[7] = block[10];
-  tmp[8] = block[17];
-  tmp[9] = block[24];
-  tmp[10] = block[32];
-  tmp[11] = block[25];
-  tmp[12] = block[18];
-  tmp[13] = block[11];
-  tmp[14] = block[4];
-  tmp[15] = block[5];
-  tmp[16] = block[12];
-  tmp[17] = block[19];
-  tmp[18] = block[26];
-  tmp[19] = block[33];
-  tmp[20] = block[40];
-  tmp[21] = block[48];
-  tmp[22] = block[41];
-  tmp[23] = block[34];
-  tmp[24] = block[27];
-  tmp[25] = block[20];
-  tmp[26] = block[13];
-  tmp[27] = block[6];
-  tmp[28] = block[7];
-  tmp[29] = block[14];
-  tmp[30] = block[21];
-  tmp[31] = block[28];
-  tmp[32] = block[35];
-  tmp[33] = block[42];
-  tmp[34] = block[49];
-  tmp[35] = block[56];
-  tmp[36] = block[57];
-  tmp[37] = block[50];
-  tmp[38] = block[43];
-  tmp[39] = block[36];
-  tmp[40] = block[29];
-  tmp[41] = block[22];
-  tmp[42] = block[15];
-  tmp[43] = block[23];
-  tmp[44] = block[30];
-  tmp[45] = block[37];
-  tmp[46] = block[44];
-  tmp[47] = block[51];
-  tmp[48] = block[58];
-  tmp[49] = block[59];
-  tmp[50] = block[52];
-  tmp[51] = block[45];
-  tmp[52] = block[38];
-  tmp[53] = block[31];
-  tmp[54] = block[39];
-  tmp[55] = block[46];
-  tmp[56] = block[53];
-  tmp[57] = block[60];
-  tmp[58] = block[61];
-  tmp[59] = block[54];
-  tmp[60] = block[47];
-  tmp[61] = block[55];
-  tmp[62] = block[62];
-  tmp[63] = block[63];
-  memcpy(block, tmp, DCTSIZE2 * sizeof(tmp[0]));
-}
-
 template <typename DI, class V>
 JXL_INLINE V NumBits(DI di, const V x) {
   // TODO(szabadka) Add faster implementations for some specific architectures.
